Rename directives file
